requirements.txt
pandas == 2.2.0
numpy == 1.26.3
matplotlib == 3.8.2
seaborn == 0.13.2
streamlit == 1.31.1
Pillow == 10.2.0
